requirements.txt
aiohttp>=3.8.5
web3>=6.0.0
python-dotenv>=0.19.0
requests>=2.31.0
numpy>=1.24.0
openpyxl>=3.1.0
tenacity>=8.2.2